# lib/api-zod/src/generated/types/publicWidgetProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicWidgetProfile {
  id: string;
  companyName: string;
  accentColor: string;
}