# lib/api-zod/src/generated/types/redesignAllowance.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RedesignAllowance {
  remaining: number;
  limit: number;
  resetsAt: Date;
  /** @nullable */
  exhaustedReason: string | null;
}