# lib/api-zod/src/generated/types/roomRedesignInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomRedesignInputRoomType } from './roomRedesignInputRoomType';

export interface RoomRedesignInput {
  /** Room photo file sent in the multipart image field. */
  image: string;
  roomType: RoomRedesignInputRoomType;
  /**
     * @minLength 3
     * @maxLength 500
     */
  designBrief: string;
}