# lib/api-zod/src/generated/types/roomRedesignResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomRedesignResultMimeType } from './roomRedesignResultMimeType';

export interface RoomRedesignResult {
  imageBase64: string;
  mimeType: RoomRedesignResultMimeType;
}