# lib/api-zod/src/generated/types/widgetRedesignInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WidgetRedesignInputRoomType } from './widgetRedesignInputRoomType';

export interface WidgetRedesignInput {
  image: Blob;
  roomType: WidgetRedesignInputRoomType;
  /**
     * @minLength 3
     * @maxLength 500
     */
  designBrief: string;
}