# lib/api-zod/src/generated/types/widgetRedesignResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WidgetRedesignResultMimeType } from './widgetRedesignResultMimeType';

export interface WidgetRedesignResult {
  imageBase64: string;
  mimeType: WidgetRedesignResultMimeType;
  checklist: string[];
  designSummary: string;
  redesignToken: string;
}