# lib/api-zod/src/generated/types/widgetRequestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WidgetRequestInputRoomType } from './widgetRequestInputRoomType';

export interface WidgetRequestInput {
  original: Blob;
  redesign: Blob;
  roomType: WidgetRequestInputRoomType;
  /**
     * @minLength 3
     * @maxLength 500
     */
  designBrief: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  homeownerName: string;
  /** @maxLength 320 */
  homeownerEmail: string;
  /** @maxLength 100 */
  homeownerPhone?: string;
  /** @maxLength 2000 */
  homeownerNotes?: string;
  redesignToken: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  requestKey: string;
}